pyproject.toml
[project]
name = "multiple-pdf-chatbot"
version = "0.1.0"
description = "A professional RAG-based chatbot for multiple PDFs"
authors = [
    {name = "Abhishek S", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.30.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-openai>=0.0.2",
    "langchain-anthropic>=0.1.0",
    "pypdf>=4.0.0",
    "faiss-cpu>=1.7.4",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.2",
    "tiktoken>=0.5.2",
    "sentence-transformers>=2.3.1",
]

[tool.poetry]
packages = [{include = "src"}]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
